Move 'stateful' tests to be called 'uncontrolled'
